hp 18.08

+++ albom_proto_2/MainWindow.xaml.cs
@@ -192,88 +192,45 @@ namespace albom_proto_2
                 string output1 = root1[0]+":"+Secret.rez+Class1.sel_dir+root2[root2.Length-1];
                 output1= output1.Replace('\\', '/');
 
+                
+                double K = Math.Ceiling(((i+1) /(Double)PhotosListBox.Items.Count) * Class1.progress_max*0.75);
+                bm.ReportProgress((int)K, 0);
 
-                double K = Convert.ToDouble(i) / PhotosListBox.Items.Count  * Class1.progress_max;
-                bm.ReportProgress((int)K, i);
 
-                using (var inputFs = new FileStream(root1, FileMode.Open, FileAccess.Read))
-                using (var outputFs = new FileStream(output1, FileMode.Create, FileAccess.Write))
+                if(root1 != output1)
                 {
-                    for (long nn=0; nn<= inputFs.Length;nn++)
+                    using (var inputFs = new FileStream(root1, FileMode.Open, FileAccess.Read))
+                    using (var outputFs = new FileStream(output1, FileMode.Create, FileAccess.Write))
                     {
-                        byte b = (byte)inputFs.ReadByte();
-                        outputFs.WriteByte(b);
-                        nn = inputFs.Position;
+                        for (long nn=0; nn<= inputFs.Length;nn++)
+                        {
+                            byte b = (byte)inputFs.ReadByte();
+                            outputFs.WriteByte(b);
+                            nn = inputFs.Position;
+                        }
                     }
+                    if (File.Exists(output1))
+                        File.Delete(root1);
                 }
-                if (File.Exists(output1))
-                File.Delete(root1);
-
-                no_met:;
+               
+                    no_met:bm.ReportProgress((int)Convert.ToDouble(i+1) / PhotosListBox.Items.Count * Class1.progress_max , 1);
             }
-
-
         }
 
         void bm_ProgressChanged(object sender, ProgressChangedEventArgs e)
         {
-            //string hh = PhotosListBox.Items[i].ToString();
-
-            //new
-            //string root = Regex.Split(PhotosListBox.Items[(int)e.UserState].ToString(), ":///")[1];
-            //string output = "";
-
-            //if (cur_met.Text.ToString() == Class1.nothing_metadata[current_languare_index])
-            //    return;
-            //else
-            //    output = root.Split(':')[0]+":"+ cur_met.Text + Class1.sel_dir + Name_file.Text;
-
             
             if (progress.Maximum == PhotosListBox.Items.Count)
-                progress.Value = PhotosListBox.SelectedIndex;
+                progress.Value = PhotosListBox.SelectedIndex+1;
             else
                 progress.Value = Convert.ToDouble(e.ProgressPercentage);
 
-            ImagesDir.Text = ffg + " (" + Convert.ToInt64(Convert.ToDouble(PhotosListBox.SelectedIndex + 1) / PhotosListBox.Items.Count * 100).ToString() + "%)";
-
-            //end
-
-            //var fileName = (string)e.Argument;
-            //var fsize = new FileInfo(root).Length;
-            //var bytesForPercent = fsize / 100;
-
-            //using (var inputFs = new FileStream(root, FileMode.Open, FileAccess.Read))
-            //using (var outputFs = new FileStream(output, FileMode.Create, FileAccess.Write))
-            //{
-            //    int counter = 0;
-            //    while (inputFs.Position < inputFs.Length)
-            //    {
-            //        byte b = (byte)inputFs.ReadByte();
-            //        outputFs.WriteByte(b);
-            //        counter++;
-
-            //        //if (counter % bytesForPercent == 0)
-            //        //    bgwCopyFile.ReportProgress(counter / (int)bytesForPercent);
-            //    }
-            //}
-
-            //if (File.Exists(destinationFileName))
-            //    File.Delete(destinationFileName);
-
-
-            //Class1.sel_img[1] = PhotosListBox.SelectedIndex.ToString();
-            //Class1.sel_img[0] = Name_file.Text;
-
-            //(Photo)PhotosListBox.SelectedItem++;
-
-
-            //if (cur_met.Text.ToString() != Class1.nothing_metadata[current_languare_index])
-            //    path_metadata.operacion(0, Class1.selecting_path + Class1.sel_dir + Name_file.Text);
+            ImagesDir.Text = ffg + " (" + Math.Ceiling(((float)e.ProgressPercentage / PhotosListBox.Items.Count) * 100) + "%)";
 
 
             long prog_val = Convert.ToInt64(Convert.ToDouble(PhotosListBox.SelectedIndex + 1) / PhotosListBox.Items.Count * 100);
 
-            PhotosListBox.SelectedIndex++;
+            if((int)e.UserState==1) PhotosListBox.SelectedIndex++;
         }
 
         void bm_RunWorkerCompleted(object sender, RunWorkerCompletedEventArgs e)
